+++ spyro/io/basicio.py
@@ -6,16 +6,18 @@ wave propagation, including file I/O for shots, receivers, and mesh data.
 
 from __future__ import with_statement
 
+import os
 import pickle
-import firedrake as fire
+import warnings
 from pathlib import Path
+
+import firedrake as fire
 import h5py
 import numpy as np
 from scipy.interpolate import griddata
-import os
-import warnings
-from .parallelism_wrappers import ensemble_save, ensemble_load
+
 from ..tools.version_control import is_firedrake_new
+from .parallelism_wrappers import ensemble_load, ensemble_save
 from .segy_io import read_segy_velocity_model
 
 if is_firedrake_new() is False:
@@ -283,6 +285,69 @@ def saving_source_and_receiver_location_in_csv(model, folder_name=None):
 
     return None
 
+
+def _read_velocity_binary3D(
+    fname,
+    nz,
+    nx,
+    ny,
+    byte_order="big",
+    axes_order=(0, 1, 2),
+    axes_order_sort="F",
+    dtype="float32",
+):
+    """Read a three-dimensional velocity model using Spyro ``(z, x, y)`` ordering.
+
+    Parameters
+    ----------
+    fname : str or pathlib.Path
+        Path to the binary velocity model.
+    nz : int
+        Number of velocity samples in the z direction.
+    nx : int
+        Number of velocity samples in the x direction.
+    ny : int
+        Number of velocity samples in the y direction.
+    byte_order : {"big", "little"}
+        Byte order of the velocity-model binary file.
+    axes_order : tuple of int
+        Permutation mapping binary axes to Spyro ``(z, x, y)`` order.
+    axes_order_sort : {"C", "F"}
+        Memory order used to reshape the binary velocity data.
+    dtype : str or numpy.dtype
+        Numeric data type stored in the velocity-model file.
+
+    Returns
+    -------
+    numpy.ndarray
+        Velocity array in canonical ``(z, x, y)`` ordering.
+    """
+    path = Path(fname)
+    if not path.exists():
+        raise FileNotFoundError(f"Velocity model not found: {path}")
+
+    if byte_order not in ("big", "little"):
+        raise ValueError("byte_order must be 'big' or 'little'.")
+    if axes_order_sort not in ("C", "F"):
+        raise ValueError("axes_order_sort must be 'C' or 'F'.")
+    if sorted(tuple(axes_order)) != [0, 1, 2]:  # noqa: C414
+        raise ValueError("axes_order must be a permutation of (0, 1, 2).")
+
+    dt = np.dtype(dtype).newbyteorder(">" if byte_order == "big" else "<")
+    raw = np.fromfile(path, dtype=dt)
+    expected = int(nz) * int(nx) * int(ny)
+    if raw.size != expected:
+        raise ValueError(
+            f"Velocity file contains {raw.size} values; expected {expected} "
+            f"for shape ({nz}, {nx}, {ny})."
+        )
+
+    canonical_shape = (int(nz), int(nx), int(ny))
+    inverse = np.argsort(np.asarray(axes_order, dtype=int))
+    raw_shape = tuple(canonical_shape[i] for i in inverse)
+    values = raw.reshape(raw_shape, order=axes_order_sort)
+    values = values.transpose(tuple(axes_order))
+    return np.asarray(values, dtype=np.float64)
 
 def _parse_axes_order(axes_order, ndim=3):
     """Convert an axis-order specification to axis names.
