+++ spyro/meshing/meshing_functions.py
@@ -554,7 +554,7 @@ class AutomaticMesh:
                 domain_xmax = box_xmax + padding_x
 
             # Interpolating sizing function from segy file
-            ef_segy2, f_min, f_max, n_samples, n_traces = create_sizing_function(  # noqa: RUF059
+            ef_segy2, _, _, n_samples, n_traces = create_sizing_function(
                 fname=fname, hmin=hmin_segy, bbox=segy_bbox, wl=wl, freq=freq,
                 pad_type=padding_type, pad_size_x=padding_x, pad_size_z=padding_z,
                 grade=grade, vp_water=vp_water
@@ -696,6 +696,7 @@ class AutomaticMesh:
         extend_segy = mesh_parameters.extend_segy
         h_padding = mesh_parameters.h_padding
         parallel = getattr(mesh_parameters, "gmsh_parallel", False)
+        gmsh_num_threads = getattr(mesh_parameters, "gmsh_num_threads", 1)
 
         byte_order = getattr(mesh_parameters, "segy_byte_order", "big")
         axes_order = tuple(
@@ -823,7 +824,6 @@ class AutomaticMesh:
                     water_interface=water_interface,
                     water_search_value=water_search_value,
                     structured_mesh=structured_mesh,
-                    minElementSize=minElementSize,
                     nz=nz,
                     nx=nx,
                     ny=ny,
@@ -857,6 +857,7 @@ class AutomaticMesh:
                     nz=nz,
                     nx=nx,
                     ny=ny,
+                    gmsh_num_threads=gmsh_num_threads,
                     comm=self.comm,
                     parallel_print=parallel_print,
                 )
